fix: rf license warnings and docs
