+++ tests/test_components/test_microwave.py
@@ -1912,17 +1912,6 @@ def test_impedance_calculator_mode_direction_handling():
     assert hasattr(impedance_mode, "values")
 
 
-def test_RF_license_suppression():
-    """Ensure license warnings are being emitted properly and default instantiations avoid the warning."""
-    original_setting = td.config.microwave.suppress_rf_license_warning
-    td.config.microwave.suppress_rf_license_warning = False
-    with AssertLogLevel("WARNING", contains_str="new license requirements"):
-        mode_spec = td.MicrowaveModeSpec()
-    with AssertLogLevel(None):
-        mode_spec = td.MicrowaveModeSpec._default_without_license_warning()
-    td.config.microwave.suppress_rf_license_warning = original_setting
-
-
 def test_microwave_mode_data_reordering_with_transmission_line_data():
     """Test that transmission_line_data is correctly reordered when modes are reordered."""
     from tidy3d.components.data.data_array import (
