+++ tidy3d/plugins/smatrix/data/terminal.py
@@ -14,7 +14,6 @@ from tidy3d.components.data.sim_data import SimulationData
 from tidy3d.components.microwave.base import MicrowaveBaseModel
 from tidy3d.components.microwave.data.monitor_data import AntennaMetricsData
 from tidy3d.constants import C_0
-from tidy3d.log import log
 from tidy3d.plugins.smatrix.component_modelers.terminal import TerminalComponentModeler
 from tidy3d.plugins.smatrix.data.base import AbstractComponentModelerData
 from tidy3d.plugins.smatrix.data.data_array import (
@@ -226,14 +225,6 @@ class TerminalComponentModelerData(AbstractComponentModelerData, MicrowaveBaseMo
     def smatrix_deembedded(self, port_shifts: np.ndarray = None) -> MicrowaveSMatrixData:
         """Interface function returns  de-embedded S-parameter matrix."""
         return self.change_port_reference_planes(self.smatrix(), port_shifts=port_shifts)
-
-    @pd.root_validator(pre=False)
-    def _warn_rf_license(cls, values):
-        log.warning(
-            "ℹ️ ⚠️ RF simulations are subject to new license requirements in the future. You have instantiated at least one RF-specific component.",
-            log_once=True,
-        )
-        return values
 
     def _monitor_data_at_port_amplitude(
         self,
